requirements.txt
requests
PyYAML
feedparser==6.0.11
requests==2.32.3
pydantic==2.8.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
markdown-it-py==3.0.0
jinja2==3.1.4
tenacity==8.5.0
tqdm==4.66.4
